requirements/augly.txt
augly
numpy
pillow
